refactor(client): update the creation form of a project to use the new form styles and the progress indicator component. Separate components by file (folder project/new/components), use TS for no complex components
