+++ client/src/project/new/components/TemplateSource.tsx
@@ -1,0 +1,56 @@
+/*!
+ * Copyright 2022 - Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ *  renku-ui
+ *
+ *  TemplateSource.tsx
+ *  Template Source field group component
+ */
+import * as React from "react";
+import { Button, ButtonGroup, FormGroup } from "reactstrap/lib";
+import { InputLabel } from "../../../utils/components/formlabels/FormLabels";
+import { NewProjectInputs } from "./newProject.d";
+
+interface TemplateSourceProps {
+  handlers: {
+    setProperty: Function;
+  };
+  input: NewProjectInputs;
+}
+
+const TemplateSource = ({ handlers, input }: TemplateSourceProps) => {
+  return (
+    <FormGroup className="field-group">
+      <InputLabel text="Template source" />
+      <br />
+      <ButtonGroup size="sm">
+        <Button color="primary" outline active={!input.userRepo}
+          onClick={(e) => handlers.setProperty("userRepo", false)}>
+          RenkuLab
+        </Button>
+        <Button color="primary" outline active={!!input.userRepo}
+          onClick={(e) => handlers.setProperty("userRepo", true)}>
+          Custom
+        </Button>
+      </ButtonGroup>
+    </FormGroup>
+  );
+};
+
+export default TemplateSource;
