+++ client/src/styles/components/_renku_form.scss
@@ -1,10 +1,3 @@
-.input-hint {
-  font-size: 14px;
-  line-height: 16px;
-  color: #626262;
-  padding-top: 5px;
-}
-
 .required-label {
   color: var(--bs-danger);
   font-size: 16px;
@@ -17,4 +10,32 @@
 
 .form-check-input:disabled {
   opacity: 0.3;
+}
+
+.field-group {
+  margin-bottom: 33px !important;
+}
+
+.form-label {
+  margin-bottom: 0;
+}
+
+.form-control.is-invalid {
+  background-image: none;
+}
+
+.form-control:focus {
+  border: 1px solid #CED4DA;
+  box-sizing: border-box;
+  box-shadow: 0 0 6px rgba(0, 0, 0, 0.1);
+}
+
+.form-control.is-invalid:focus {
+  border: 1px solid var(--bs-danger);
+  box-sizing: border-box;
+  box-shadow: 0 0 6px var(--bs-rk-danger-shadow);
+}
+
+.field-group {
+  margin-bottom: 33px;
 }
