--- client/src/utils/components/FieldGroups.js
@@ -1,49 +1,0 @@
-/*!
- * Copyright 2022 - Swiss Data Science Center (SDSC)
- * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
- * Eidgenössische Technische Hochschule Zürich (ETHZ).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- *  renku-ui
- *
- *  FieldGroup.js
- *  FieldGroup component.
- */
-
-import React, { Component } from "react";
-import {
-  FormFeedback, FormGroup, FormText, Input, Label
-} from "reactstrap";
-
-class FieldGroup extends Component {
-  render() {
-    const label = this.props.label,
-      help = this.props.help,
-      feedback = this.props.feedback,
-      props = this.props;
-    const subProps = {};
-    if (props.valid === true)
-      subProps.valid = "true";
-    if (props.invalid === true)
-      subProps.invalid = "true";
-    return <FormGroup>
-      <Label>{label}</Label>
-      <Input {...props} />
-      {feedback && <FormFeedback {...subProps}>{feedback}</FormFeedback>}
-      {help && <FormText color="muted">{help}</FormText>}
-    </FormGroup>;
-  }
-}
-export { FieldGroup };
