+++ client/src/utils/components/formlabels/FormLabels.tsx
@@ -1,3 +1,27 @@
+/*!
+ * Copyright 2022 - Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ *  renku-ui
+ *
+ *  FormLabels.tsx
+ *  FormLabels components.
+ */
 import * as React from "react";
 import { FormText, FormFeedback, Label } from "reactstrap/lib";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
@@ -8,10 +32,11 @@ import { Loader } from "../Loader";
 
 interface LabelProps {
   text: string;
+  children?: React.ReactNode
 }
 
 interface InputLabelProps extends LabelProps {
-  isRequired: boolean;
+  isRequired?: boolean;
   isOptional?: boolean;
 }
 
@@ -47,10 +72,10 @@ const InputHintLabel = ({ text }: LabelProps) => {
   );
 };
 
-const ErrorLabel = ({ text }: LabelProps) => {
+const ErrorLabel = ({ text, children }: LabelProps) => {
   return (
     <FormFeedback className="error-feedback">
-      <FontAwesomeIcon icon={faExclamationTriangle} />{" "}{ text }
+      <FontAwesomeIcon icon={faExclamationTriangle} />{" "}{ text } {children}
     </FormFeedback>);
 };
